Noisy Actions & Noise-Masked Observations (#14)
* working version of noisy adversary

* remove unused import

* fix line length

* Actually commit linting changes

* Add additive noise to observation masking code

* clean up score file'

* update for running ray

* linting fixes

* manually specify best adversaries path

* tiny readme change

* Working local version of experiments

* linting

* clean up experiment ordering, make medium accuracy config right number of episodes

* Update score agent code to keep track of more keys in output JSON index

* fix lognormal and episodes and envs

* Update to comment out masking and turn up episodes

* only run noisy adversary actions

* hacky solution to add noise magnitude to index

* up number of episodes and samples

* Workaround pip issue #2195

* comment out to match remote experiments being run

* only run noise victim obs against adversary with reasonable episodes and samples

* update number of episodes

* fix accidental duplicate observer

* pre-linting commit

* attempts to get ray aws to work

* failing s3 command location

* reduce num envs, increase episodes and samples

* config changes

* local vs amazon changes

* moderate changes to score

* Fix AWS config

* increase envs and create smaller noise config

* small remote changes

* Quicker defaults for Cody

* local changes to increase episodes

* train with all Zoo

* back to quick envs

* switch back to all Zoo envs

* add noisy observation plotting

* clean up accidental README commit

* uncomment experiment file

* clean up out_root for experiment file

* clean up last minute changes to configs-multi-score

* separate and clean up ray config files

* undo short term hack to make copying work

* remove unused parts of experiment script

* Combine adversary_path lines

* remove incorrect comment

* fix issue with paths overriding one another

* Fix score.copy()

* update experiment name

* turn up target utilization

* fix sync command to get dot files and remove whitespac

* remove mask_agent_kwargs and pull out masking_type to be a top level parameter

* change paths to reference /data/aws

* factor out wait_prof into common.sh

* Clean up bash scripts

* Reduce code duplication in noisy observation plotting

* Refactor code out of common

* Factor out into policies NoisyAgentWrapper that score_agent depends on

* Bugfix in score_agent

* add command line arguments for plotting noisy observations by implementing sacred

* pep8-fix multi line string

* add EC2 role to aws.yaml

* Revert unnecessary aws.yaml change

* Reduce code duplication in configs.multi.score

* Simplify noisy_actions_and_obs.sh; delete extra Ray AWS config, include sync instructions in bash script

* Fix errors introduced by automatic policy code moval

* Dead code removal in visualize.noisy_observations

* Add new line

* Fix typo in multi.score

* Add test cases
