+++ experiments/modelfree/attack_transfer.sh
@@ -1,16 +1,8 @@
 #!/usr/bin/env bash
 
-function wait_proc {
-    if [[ -f ~/ray_bootstrap_config.yaml ]]; then
-        # Running on a Ray cluster. We want to submit all the jobs in parallel.
-        sleep 5  # stagger jobs a bit
-    else
-        # Running locally. Each job will start a Ray cluster. Submit sequentially.
-        wait
-    fi
-}
-
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
+. ${DIR}/common.sh
+
 OUT_ROOT=data/aws/score_agents
 TIMESTAMP=`date --iso-8601=seconds`
