+++ experiments/modelfree/baselines.sh
@@ -1,14 +1,7 @@
 #!/usr/bin/env bash
 
-function wait_proc {
-    if [[ -f ~/ray_bootstrap_config.yaml ]]; then
-        # Running on a Ray cluster. We want to submit all the jobs in parallel.
-        sleep 5  # stagger jobs a bit
-    else
-        # Running locally. Each job will start a Ray cluster. Submit sequentially.
-        wait
-    fi
-}
+DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
+. ${DIR}/common.sh
 
 CMD="python -m modelfree.multi.score with high_accuracy"
 OUT_DIR=data/aws/score_agents
