+++ requirements.txt
@@ -17,3 +17,5 @@ codecov>=2.0.15
 pytest-cov>=2.6.1
 flake8>=3.7
 isort>=4.3
+statsmodels>=0.9.0
+seaborn>=0.9.0
