+++ src/modelfree/configs/ray/baremetal.yaml
@@ -38,7 +38,8 @@ head_start_ray_commands:
       git remote set-url origin https://github.com/HumanCompatibleAI/adversarial-policies.git &&
       git pull &&
       env=modelfree . /adversarial-policies/ci/prepare_env.sh &&
-      pip install . &&
+      python3 setup.py sdist bdist_wheel &&
+      pip install dist/aprl-*.whl &&
       ray stop &&
       ulimit -n 65536 &&
       ray start --head --redis-port=6379 --object-manager-port=8076 --autoscaling-config=~/ray_bootstrap_config.yaml
@@ -47,7 +48,8 @@ worker_start_ray_commands:
       git remote set-url origin https://github.com/HumanCompatibleAI/adversarial-policies.git &&
       git pull &&
       env=modelfree . /adversarial-policies/ci/prepare_env.sh &&
-      pip install . &&
+      python3 setup.py sdist bdist_wheel &&
+      pip install dist/aprl-*.whl &&
       ray stop &&
       ulimit -n 65536 &&
       ray start --redis-address=$RAY_HEAD_IP:6379 --object-manager-port=8076
