+++ src/modelfree/policies/wrappers.py
@@ -1,0 +1,38 @@
+import numpy as np
+
+from modelfree.policies.base import DummyModel
+
+
+class NoisyAgentWrapper(DummyModel):
+    def __init__(self, agent, noise_annealer, noise_type='gaussian'):
+        """
+        Wrap an agent and add noise to its actions
+        :param agent: (BaseRLModel) the agent to wrap
+        :param noise_annealer: Annealer.get_value - presumably the noise should be decreased
+        over time in order to get the adversarial policy to perform well on a normal victim.
+        :param noise_type: str - the type of noise parametrized by noise_annealer's value.
+        Current options are [gaussian]
+        """
+        super().__init__(policy=agent, sess=agent.sess)
+        self.noise_annealer = noise_annealer
+        self.noise_generator = self._get_noise_generator(noise_type)
+
+    @staticmethod
+    def _get_noise_generator(noise_type):
+        noise_generators = {
+            'gaussian': lambda x, size: np.random.normal(scale=x, size=size)
+        }
+        return noise_generators[noise_type]
+
+    def log_callback(self, logger):
+        current_noise_param = self.noise_annealer()
+        logger.logkv('shaping/victim_noise', current_noise_param)
+
+    def predict(self, observation, state=None, mask=None, deterministic=False):
+        original_actions, states = self.policy.predict(observation, state, mask, deterministic)
+        action_shape = original_actions.shape
+        noise_param = self.noise_annealer()
+
+        noise = self.noise_generator(noise_param, action_shape)
+        noisy_actions = original_actions * (1 + noise)
+        return noisy_actions, states
